+++ Example/Source/API/Clients/Client.swift
@@ -14,9 +14,7 @@ protocol Targetable {
     associatedtype Target: Endpoint
 }
 
-class Client<T: Endpoint>: Leash.Client, Targetable {
-    
-    typealias Target = T
+class Client<Target: Endpoint>: Leash.Client, Targetable {
     
     override func urlRequest(for endpoint: Leash.Endpoint) throws -> URLRequest {
         var request = try super.urlRequest(for: endpoint)
@@ -29,16 +27,45 @@ class Client<T: Endpoint>: Leash.Client, Targetable {
 extension Client {
     
     @discardableResult
-    func execute<T: Decodable>(_ endpoint: Target, completion: @escaping (Response<T>) -> ()) -> DataRequest? {
+    func execute<T: Decodable>(_ endpoint: Target, completion: @escaping APICompletion<T>) -> DataRequest? {
         return execute(endpoint as Leash.Endpoint, completion: completion)
     }
     
 }
 
-extension Reactive where Base: Leash.Client & Targetable {
+extension Reactive where Base: Leash.Client {
     
-    func execute<T: Decodable>(_ endpoint: Base.Target) -> Single<T> {
-        return execute(endpoint as Leash.Endpoint)
+    func execute<T: Decodable>(_ execution: @escaping (@escaping APICompletion<T>) -> ()) -> Single<T> {
+        return Single.create { single in
+            execution { response in
+                single(SingleEvent.fromResponse(response))
+            }
+            
+            return Disposables.create { }
+        }
+    }
+    
+    func execute<T: Decodable, U>(_ execution: @escaping (U, @escaping APICompletion<T>) -> (), with request: U) -> Single<T> {
+        return Single.create { single in
+            execution(request) { response in
+                single(SingleEvent.fromResponse(response))
+            }
+            
+            return Disposables.create { }
+        }
+    }
+    
+}
+
+private extension SingleEvent {
+    
+    static func fromResponse(_ response: APIResponse<Element>) -> SingleEvent<Element> {
+        switch response {
+        case .success(let value):
+            return .success(value)
+        case .failure(let error):
+            return .error(error)
+        }
     }
     
 }