+++ Source/Core/Interceptors/SerializationInterceptor.swift
@@ -1,7 +1,7 @@
 //
 //  SerializationInterceptor.swift
 //
-//  Copyright (c) 2017-2019 Luciano Polit <lucianopolit@gmail.com>
+//  Copyright (c) 2017-2020 Luciano Polit <lucianopolit@gmail.com>
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -29,6 +29,6 @@ public protocol SerializationInterceptor: Interceptor {
     /// Main method responsible for handling the interception.
     func intercept<T: DataResponseSerializerProtocol>(chain: InterceptorChain<T.SerializedObject>,
                                                       response: Response<Data>,
-                                                      result: Result<T.SerializedObject>,
+                                                      result: Result<T.SerializedObject, Swift.Error>,
                                                       serializer: T)
 }