+++ Source/RxSwift/Rx+Client.swift
@@ -1,7 +1,7 @@
 //
 //  Rx+Client.swift
 //
-//  Copyright (c) 2017-2019 Luciano Polit <lucianopolit@gmail.com>
+//  Copyright (c) 2017-2020 Luciano Polit <lucianopolit@gmail.com>
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -25,8 +25,6 @@
 import Foundation
 import RxSwift
 
-// MARK: - Reactive
-
 extension Client: ReactiveCompatible { }
 
 extension Reactive where Base: Client {
@@ -41,7 +39,13 @@ extension Reactive where Base: Client {
     public func execute<T: Decodable>(_ endpoint: Endpoint,
                                       type: ReactiveResponse<T>.Type = ReactiveResponse<T>.self,
                                       queue: DispatchQueue? = nil) -> Single<ReactiveResponse<T>> {
-        return base.execute(endpoint, queue: queue)
+        return Single.create { single in
+            self.base.execute(endpoint, queue: queue) { response in
+                single(SingleEvent.fromResponse(response))
+            }
+            
+            return Disposables.create { }
+        }
     }
     
     /// Creates and executes the request for the specified endpoint.
@@ -54,22 +58,7 @@ extension Reactive where Base: Client {
     public func execute<T: Decodable>(_ endpoint: Endpoint,
                                       type: T.Type = T.self,
                                       queue: DispatchQueue? = nil) -> Single<T> {
-        return base.execute(endpoint, queue: queue).justValue()
-    }
-    
-}
-
-// MARK: - Utils
-
-private extension Client {
-    
-    func execute<T: Decodable>(_ endpoint: Endpoint, queue: DispatchQueue? = nil) -> Single<ReactiveResponse<T>> {
-        do {
-            let request = try self.request(for: endpoint)
-            return request.rx.responseDecodable(queue: queue, client: self, endpoint: endpoint)
-        } catch {
-            return .error(Error.encoding(error))
-        }
+        return execute(endpoint, queue: queue).justValue()
     }
     
 }