+++ Tests/RxSwift/Rx+Alamofire+Leash+Tests.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import Alamofire
 import RxSwift
 import XCTest
 import OHHTTPStubs
@@ -76,7 +77,7 @@ extension AlamofireLeashTests {
         assertNoErrorThrown {
             let endpoint = successEndpoint
             let dataRequest = try client.request(for: endpoint)
-            let serializer = DataRequest.jsonResponseSerializer(options: .allowFragments)
+            let serializer = JSONResponseSerializer()
             dataRequest.rx.response(client: client, endpoint: endpoint, serializer: serializer)
                 .subscribe(onSuccess: { _ in
                     expectation.fulfill()
@@ -108,7 +109,7 @@ extension AlamofireLeashTests {
         let disposeBag = DisposeBag()
         let undecodableEndpoint = Endpoint(path: "decoding/error")
         stub(condition: isEndpoint(undecodableEndpoint)) { _ in
-            return OHHTTPStubsResponse(data: Data(), statusCode: 200, headers: nil)
+            return OHHTTPStubsResponse(jsonObject: ["some": "some"], statusCode: 200, headers: nil)
         }
         rx.executeRequest(builder: builder, endpoint: undecodableEndpoint)?
             .subscribe(onError: { error in